# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Health Triage Assistant API",
    "description": "### Offline-First Health Triage Assistant API\n\nProduction-grade RESTful API serving clinical triage evaluation, emergency dispatch, health profile management, and PWA outbox sync.\n\n**Key Architectural Highlights:**\n- **Dual Intelligence**: Deterministic rule engine + optional Gemini AI enhancement\n- **Multilingual**: BCP 47 language tagging (English `en`, Twi `tw`)\n- **Clean Architecture**: Isolated domain models, services, and interface adapters\n- **RFC 7807 Errors**: Structured problem details error payloads\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user account",
        "description": "Creates a new user account with phone number and password. Returns a JWT access token on success.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authenticate and receive JWT token",
        "description": "Validates phone number and password credentials. Returns a JWT access token on successful authentication.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessment/start": {
      "post": {
        "tags": [
          "Assessment API"
        ],
        "summary": "Create a new assessment session",
        "description": "Initializes a new interactive assessment session.",
        "operationId": "start_assessment_api_v1_assessment_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/assessment/symptoms": {
      "post": {
        "tags": [
          "Assessment API"
        ],
        "summary": "Set primary symptom for assessment",
        "description": "Accepts primary symptom selection and returns the first follow-up question.",
        "operationId": "set_symptoms_api_v1_assessment_symptoms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentSymptomsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentSymptomsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessment/answer": {
      "post": {
        "tags": [
          "Assessment API"
        ],
        "summary": "Submit question answer",
        "description": "Records a question response and returns the next question or final result.",
        "operationId": "submit_answer_api_v1_assessment_answer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentAnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessment/{session_id}": {
      "get": {
        "tags": [
          "Assessment API"
        ],
        "summary": "Retrieve assessment progress",
        "description": "Gets the current status and answer count of an assessment session.",
        "operationId": "get_assessment_progress_api_v1_assessment__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessment/{session_id}/result": {
      "get": {
        "tags": [
          "Assessment API"
        ],
        "summary": "Retrieve assessment result",
        "description": "Gets the evaluated severity, recommendations, and emergency status for a session.",
        "operationId": "get_assessment_result_api_v1_assessment__session_id__result_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessment/restart": {
      "post": {
        "tags": [
          "Assessment API"
        ],
        "summary": "Restart an assessment",
        "description": "Marks the previous session as abandoned and creates a fresh session.",
        "operationId": "restart_assessment_api_v1_assessment_restart_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentRestartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentRestartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssessmentAnswerRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "node_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Node Id"
          },
          "answer_value": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Answer Value"
          },
          "answer_raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Raw Text"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "node_id",
          "answer_value"
        ],
        "title": "AssessmentAnswerRequest",
        "description": "Payload to record an answer to a question node."
      },
      "AssessmentAnswerResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "is_completed": {
            "type": "boolean",
            "title": "Is Completed"
          },
          "next_question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NextQuestionDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentResultResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "is_completed"
        ],
        "title": "AssessmentAnswerResponse",
        "description": "Response returned after answering a question."
      },
      "AssessmentProgressResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "symptom_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symptom Id"
          },
          "answers_count": {
            "type": "integer",
            "title": "Answers Count"
          },
          "conducted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Conducted At"
          },
          "created_offline": {
            "type": "boolean",
            "title": "Created Offline"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "answers_count",
          "conducted_at",
          "created_offline"
        ],
        "title": "AssessmentProgressResponse",
        "description": "Response representing current session progress."
      },
      "AssessmentRestartRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "AssessmentRestartRequest",
        "description": "Payload to restart an existing assessment session."
      },
      "AssessmentRestartResponse": {
        "properties": {
          "new_session_id": {
            "type": "string",
            "title": "New Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Assessment session restarted successfully."
          }
        },
        "type": "object",
        "required": [
          "new_session_id",
          "status"
        ],
        "title": "AssessmentRestartResponse",
        "description": "Response returned upon restarting an assessment session."
      },
      "AssessmentResultResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "severity": {
            "$ref": "#/components/schemas/UrgencyCode"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "is_emergency": {
            "type": "boolean",
            "title": "Is Emergency"
          },
          "conducted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Conducted At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "severity",
          "recommendations",
          "explanation",
          "is_emergency",
          "conducted_at"
        ],
        "title": "AssessmentResultResponse",
        "description": "Final assessment result model."
      },
      "AssessmentStartRequest": {
        "properties": {
          "language_code": {
            "type": "string",
            "pattern": "^(en|tw)$",
            "title": "Language Code",
            "default": "en"
          },
          "consultation_mode": {
            "$ref": "#/components/schemas/ConsultationMode",
            "default": "TEXT"
          },
          "created_offline": {
            "type": "boolean",
            "title": "Created Offline",
            "default": false
          }
        },
        "type": "object",
        "title": "AssessmentStartRequest",
        "description": "Payload to start a new assessment session."
      },
      "AssessmentStartResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "language_code": {
            "type": "string",
            "title": "Language Code"
          },
          "consultation_mode": {
            "$ref": "#/components/schemas/ConsultationMode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "language_code",
          "consultation_mode",
          "created_at"
        ],
        "title": "AssessmentStartResponse",
        "description": "Response returned upon starting a session."
      },
      "AssessmentSymptomsRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "symptom_slug": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Symptom Slug"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "symptom_slug"
        ],
        "title": "AssessmentSymptomsRequest",
        "description": "Payload to set the initial primary symptom for a session."
      },
      "AssessmentSymptomsResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "symptom_slug": {
            "type": "string",
            "title": "Symptom Slug"
          },
          "next_question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NextQuestionDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_emergency": {
            "type": "boolean",
            "title": "Is Emergency",
            "default": false
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UrgencyCode"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "symptom_slug"
        ],
        "title": "AssessmentSymptomsResponse",
        "description": "Response returned after setting primary symptom."
      },
      "ConsultationMode": {
        "type": "string",
        "enum": [
          "TEXT",
          "VOICE",
          "HYBRID"
        ],
        "title": "ConsultationMode",
        "description": "How the user interacted with the triage flow."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "phone_number": {
            "type": "string",
            "maxLength": 20,
            "minLength": 7,
            "title": "Phone Number"
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "phone_number",
          "password"
        ],
        "title": "LoginRequest",
        "description": "User login request payload."
      },
...
      "RegisterRequest": {
        "properties": {
          "phone_number": {
            "type": "string",
            "maxLength": 20,
            "minLength": 7,
            "title": "Phone Number",
            "description": "Phone number in E.164 format (e.g. +233241234567)",
            "examples": [
              "+233241234567"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 8,
            "title": "Password"
          },
          "preferred_language": {
            "type": "string",
            "pattern": "^(en|tw)$",
            "title": "Preferred Language",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "phone_number",
          "password"
        ],
        "title": "RegisterRequest",
        "description": "User registration request payload."
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in_minutes": {
            "type": "integer",
            "title": "Expires In Minutes"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in_minutes"
        ],
        "title": "TokenResponse",
        "description": "JWT access token response payload."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}